requirements.txt
pandas==3.0.1
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2